requirements.txt
streamlit==1.25.0
speechrecognition==3.9.0
python-dotenv==1.0.0
deepl==1.13.1
edge-tts==1.3.5
